64-bit fixes and windows 64-bit libraries. Dll files are provided as separate download.

+++ source/voxelman/server/chunkman.d
@@ -192,7 +192,7 @@ struct ChunkMan
 		assert(clientInfo, "clientStorage[clientId] is null");
 		ChunkRange oldRegion = clientInfo.visibleRegion;
 		vec3 cameraPos = clientInfo.pos;
-		size_t viewRadius = clientInfo.viewRadius;
+		int viewRadius = clientInfo.viewRadius;
 
 		ivec3 chunkPos = cameraToChunkPos(cameraPos);
 		ChunkRange newRegion = calcChunkRange(chunkPos, viewRadius);
